Exit early with useful error message when 'build_failed' is true
Adds a new input 'build_failed' which, when true, causes the deploy action
to exit early with a useful error message. Without this, when a build fails
and the deploy step runs after it, the error message refers to a build
artifact not being found, and can confuse a reader.